+++ code/system/sys.fs/src/MemoryMock/Mock.Indexer.mts
