+++ code/system/sys.fs/src/MemoryMock/common.mts
