+++ code/system/sys.fs/src/MemoryMock/index.mts
