+++ code/system/sys.fs/src/MemoryMock/util.mts
