+++ code/system/sys.fs/src/TEST/index.mts
@@ -3,5 +3,5 @@ export { expect, expectError } from 'sys.test';
 
 export * from '../common/index.mjs';
 export { Filesystem } from '../index.mjs';
-export { FsMock } from '../Mock/index.mjs';
+export { FsMock } from '../MemoryMock/index.mjs';
 export { TestPrep } from './TestPrep.mjs';
